componentize range picker
