+++ src/components/ChartControls.vue
@@ -8,30 +8,17 @@
       @click="select(index)">
       {{ button.label }}
     </app-chip>
-    <!--    todo componentize-->
-    <div class="b">
-      <VueDatePicker
-        ref="datepicker"
-        v-model="date"
-        range
-        :clearable="false"
-        auto-apply
-        hide-input-icon
-        :enable-time-picker="false"
-        :partial-range="false"
-        :max-date="today"
-        placeholder="CUSTOM"
-        :input-class-name="`dp-custom-input ${selectedIndex === 'custom' ? 'dp-custom-active' : ''}`"
-        @range-end="setActive"
-        @update:model-value="doIt"/>
-    </div>
+    <range-picker
+      :selected-index="selectedIndex"
+      :is-range-set="hasCustomDate"
+      @updated="fetchIt"
+      @activated="setActive"/>
   </div>
 </template>
 
 <script setup>
-import VueDatePicker from '@vuepic/vue-datepicker'
-import '@vuepic/vue-datepicker/dist/main.css'
-import { DateTime } from 'luxon'
+
+import RangePicker from '~/components/RangePicker'
 
 const buttons = [
   { interval: 'day', limit: '7', label: '1W' },
@@ -40,98 +27,29 @@ const buttons = [
   { interval: 'month', limit: '12', label: '1Y' },
   { interval: 'month', limit: '100', label: 'ALL' },
 ]
-const date = ref()
-const datepicker = ref(null)
-
-// todo limit at beginning
-const from = ref(DateTime.now().minus({ days: 1 }).toFormat('yyyy-MM-dd'))
-const to = ref(DateTime.now().toFormat('yyyy-MM-dd'))
-const today = DateTime.now().toFormat('yyyy-MM-dd')
 
 const selectedIndex = ref(0)
+const hasCustomDate = ref(true)
 
-function doIt(param) {
-  console.log('param', param)
+function select(value) {
+  selectedIndex.value = value
+  emit('selected', buttons[value])
+  hasCustomDate.value = false
 }
 
 function setActive() {
+  hasCustomDate.value = true
   selectedIndex.value = 'custom'
 }
 
-function select(value) {
-  closeDatepicker()
-  selectedIndex.value = value
-  emit('selected', buttons[value])
-}
-
-function closeDatepicker() {
-  date.value = null
-  if (datepicker) {
-    datepicker.value.closeMenu()
-  }
+function fetchIt(dateRange) {
+  console.log('fetchIt', dateRange)
 }
 
 const emit = defineEmits(['selected'])
 </script>
 
-<style>
-/*todo classes*/
-.dp__main {
-  display: inline-flex;
-}
-
-.dp__main > div {
-  width: 100%;
-}
-
-.b {
-  grid-column: span 5;
-
-}
-
-.dp-custom-input {
-  border-radius: 4px;
-  font-family: var(--font-monospaced);
-  align-items: center;
-  text-align: center;
-  height: 28px;
-  font-size: 12px;
-  line-height: 20px;
-  padding: var(--space-0) var(--space-1);
-  background: var(--color-snow);
-  color: var(--color-midnight);
-  border: 0;
-  font-weight: 400;
-
-  width: 100%;
-  @media (--desktop) {
-    width: 68px;
-
-    height: 32px;
-    font-size: 14px;
-    padding: var(--space-0) var(--space-1);
-  }
-}
-
-.dp__range_start {
-  background: var(--color-error) !important;
-  color: var(--color-white);
-}
-
-.dp__range_end {
-  background: var(--color-error) !important;
-  color: var(--color-white);
-}
-
-.dp-custom-active {
-  background: var(--color-error) !important;
-  color: var(--color-white);
-  width: 100%;
-  @media (--desktop) {
-    width: 208px;
-  }
-}
-
+<style scoped>
 .chart-controls {
   display: grid;
   grid-template-columns: 1fr 1fr 1fr 1fr 1fr;
